pyproject.toml
[tool.poetry]
name = "weather-app"
version = "0.1.0"
description = "A web app made by django that displays the weather of any given city."
authors = ["samuel woldegiorgies"]
readme = "README.md"
packages = [{include = "weatherapp"}, {include = "weather"}, {include = "users"}]

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
Django = "^4.0.1"
gunicorn = "^20.1.0"
psycopg2 = "^2.9.3"
python-dotenv = "^0.19.2"
requests = "^2.26.0"
whitenoise = "^5.3.0"
dj-database-url = "^0.5.0"
asgiref = "^3.4.1"
certifi = "^2021.10.8"
charset-normalizer = "^2.0.9"
idna = "^3.3"
sqlparse = "^0.4.2"
urllib3 = "^1.26.7"
sentry-sdk = "^2.42.0"

[tool.poetry.group.dev.dependencies]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
manage = "manage:main"
